Move UsageType and SynchronizationType to embassy_usb_driver
