+++ embassy-usb/src/descriptor.rs
@@ -2,7 +2,7 @@
 use embassy_usb_driver::EndpointType;
 
 use crate::builder::Config;
-use crate::driver::EndpointInfo;
+use crate::driver::{EndpointInfo, SynchronizationType, UsageType};
 use crate::types::{InterfaceNumber, StringIndex};
 use crate::CONFIGURATION_VALUE;
 
@@ -37,40 +37,6 @@ pub mod capability_type {
     pub const SS_USB_DEVICE: u8 = 3;
     pub const CONTAINER_ID: u8 = 4;
     pub const PLATFORM: u8 = 5;
-}
-
-/// USB endpoint synchronization type. The values of this enum can be directly
-/// cast into `u8` to get the bmAttributes synchronization type bits.
-/// Values other than `NoSynchronization` are only allowed on isochronous endpoints.
-#[repr(u8)]
-#[derive(Copy, Clone, Eq, PartialEq, Debug)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum SynchronizationType {
-    /// No synchronization is used.
-    NoSynchronization = 0b00,
-    /// Unsynchronized, although sinks provide data rate feedback.
-    Asynchronous = 0b01,
-    /// Synchronized using feedback or feedforward data rate information.
-    Adaptive = 0b10,
-    /// Synchronized to the USB’s SOF.
-    Synchronous = 0b11,
-}
-
-/// USB endpoint usage type. The values of this enum can be directly cast into
-/// `u8` to get the bmAttributes usage type bits.
-/// Values other than `DataEndpoint` are only allowed on isochronous endpoints.
-#[repr(u8)]
-#[derive(Copy, Clone, Eq, PartialEq, Debug)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum UsageType {
-    /// Use the endpoint for regular data transfer.
-    DataEndpoint = 0b00,
-    /// Endpoint conveys explicit feedback information for one or more data endpoints.
-    FeedbackEndpoint = 0b01,
-    /// A data endpoint that also serves as an implicit feedback endpoint for one or more data endpoints.
-    ImplicitFeedbackDataEndpoint = 0b10,
-    /// Reserved usage type.
-    Reserved = 0b11,
 }
 
 /// A writer for USB descriptors.
